Safari-611.1.10/LayoutTests/fast/regex/quantified-assertions-expected.txt

This page tests assertions followed by quantifiers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Testing regexp: /(?=a){0}/gm
PASS regexp.test('a') is true
PASS regexp.lastIndex is 0

Testing regexp: /(?=a){1}/gm
PASS regexp.test('a') is true
PASS regexp.lastIndex is 0

Testing regexp: /(?!a){0}/gm
PASS regexp.test('b') is true
PASS regexp.lastIndex is 0

Testing regexp: /(?!a){1}/gm
PASS regexp.test('b') is true
PASS regexp.lastIndex is 0
PASS /^(?=a)?b$/.test("b") is true
PASS successfullyParsed is true

TEST COMPLETE
